refactor: utils 폴더를 infra로 리네임
- src/utils → src/infra
- tests/utils → tests/infra
- 인프라 관련 코드임을 명확히 표현
