+++ src/main.py
@@ -4,11 +4,11 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI
 from fastapi.staticfiles import StaticFiles
 
+from src.infra.redis import close_redis
+from src.infra.storage import get_storage
+from src.infra.storage.local import LocalStorage
 from src.routes.job import router as job_router
 from src.routes.upload import router as upload_router
-from src.utils.redis import close_redis
-from src.utils.storage import get_storage
-from src.utils.storage.local import LocalStorage
 
 
 @asynccontextmanager
